PHR: separate ArpTable and utils into mods
